bug fix: 'bunch' class template had int for channel count, now size_t: this was a relic from the port from vspline. My compilers on debian12 did accept the int in this place, but on my mac, the compilers both rejected the code.

+++ zimt/eval.h
@@ -1745,7 +1745,7 @@ private:
   /// 'synthetic' types, using the templates 'bunch' and 'vector'.
   /// The reinterpreted data are fed to 'inner'.
 
-  template < template < typename , int > class bunch ,
+  template < template < typename , size_t > class bunch ,
              template < typename , size_t > class vector ,
              size_t VSZ ,
              typename in_type ,
